requirements.txt
bs4
utils
jmespath
requests
sqlalchemy
shapely
requests-ftp
enum34
python-dateutil
monthdelta
pyproj==1.9.6
OWSLib==0.16
stringcase
boto3
xmltodict